Apply template: Fix merge options for markdown text fields
The markdown rich text fields shadow the underlying textarea. Thus, the
update of the textarea's value has no effect.
This commit stores the MDE JavaScript objects on the the field's DOM
object. The change of the contents is now performed on the MDE elements
using their designated method rather than directly on the textarea
element.