integration-tests: add wait/retry when checking for duplicates (#2101)
Quite often the Travis build fails because the integration tests sees an incorrect amount of duplicates. But sometimes the exact same commit builds fine. The theory is that Travis is sometimes so slow that it takes a bit of time for the background celery task to run/complete.

This PR implements a 5s wait time and 4 tries when checking for duplicates and hopefully makes the builds more stable.