set mitigation related fields correctly at close_old_findings (#3929)
* Bugfix at close_old_findings

Fixed a bug where the 'is_Mitigated' field is not updated when findings are automatically closed, which leads to them not been considered closed internally by DefectDojo (although on the UI it appears "Mitigated"). This causes these findings not to appear on metrics nor at the 'View Closed Findings' option.

* Minor refactoring at close_old_findings

The logic to update mitigated and mitigated_by was already written at dojo/finding/helper.py#update_finding_status, so we're using that to avoid code duplication.