Make SLA calculation of a Finding Group use the most severe finding (#7102)
When using finding groups within a test, the SLA calculation for groups passes the entire group to the SLA calculation for a single finding. There is an expected mismatch there. This PR takes the most severe finding I the group and passes that finding the SLA calculation function so that the group has the fastest approaching SLA.